requirements.txt
aws-cdk-lib
constructs
requests
python-dotenv
